Outline with width working
It is not the prettiest of the solutions, it is fast and better than the one in the presentation

I have been tinkering with trying to make it procedurally bigger but it costs a lot of gpu cycles although there was no branching, this ones starts leaving holes in sharp obtuse corners which would require more steps which would not be that hard to add but a little painful that's it.

This is enough and there is alpha and with controls.