Depth issue solved + Take out bad reprojection of the world position
Was reprojecting to "frustrum" coordinates when it shouldn't
Added the proper depth calculations to write

BIG ISSUE - Fargems become corrupted in chunks when not in the *grid domain, (y > 0.) then if you look there it stays corrupted. I don't know what is causing it
Maybe it is caused by not using specific textures or not specifying drawbuffers, but then it should also be fucked on the main fboColor when I did not do that.