blur was being done incorrectly
I was only really writing 1 pass to each texture instead of both to a texture
Maybe I should do a uniform for expected passed, add a blend form so that it keeps being being added or something?

I don't want to end up doing the blur in a single pass, dual pass is the way technically
But I don't  know again how to setup the textures for read after a write.